bug fix for arbitrary work dir
